Improve compatibility of api/issues/search2
