+++ server/sonar-server/src/test/resources/org/sonar/server/issue/ws/SearchActionMediumTest/issue_on_removed_file.json
@@ -13,12 +13,10 @@
   ],
   "components": [
     {
-      "id": "REMOVED_FILE_ID",
       "key": "REMOVED_FILE_KEY",
       "enabled" : false
     },
     {
-      "id": "PROJECT_ID",
       "key": "PROJECT_KEY",
       "enabled" : true
     }
